# Make better plan when join relation's locus is general