--- a/src/test/regress/expected/join.out
+++ b/src/test/regress/expected/join.out
@@ -3585,11 +3585,11 @@ left join
 using (join_key);
                             QUERY PLAN                            
 ------------------------------------------------------------------
- Hash Right Join
+ Gather Motion 3:1  (slice3; segments: 3)
    Output: "*VALUES*".column1, i1.f1, (666)
-   Hash Cond: (i1.f1 = "*VALUES*".column1)
-   ->  Gather Motion 3:1  (slice2; segments: 3)
-         Output: i1.f1, (666)
+   ->  Hash Right Join
+         Output: "*VALUES*".column1, i1.f1, (666)
+         Hash Cond: (i1.f1 = "*VALUES*".column1)
          ->  Hash Right Join
                Output: i1.f1, 666
                Hash Cond: (i2.unique2 = i1.f1)
@@ -3602,13 +3602,15 @@ using (join_key);
                      Output: i1.f1
                      ->  Seq Scan on public.int4_tbl i1
                            Output: i1.f1
-   ->  Hash
-         Output: "*VALUES*".column1
-         ->  Values Scan on "*VALUES*"
+         ->  Hash
                Output: "*VALUES*".column1
+               ->  Redistribute Motion 1:3  (slice2; segments: 1)
+                     Output: "*VALUES*".column1
+                     Hash Key: "*VALUES*".column1
+                     ->  Values Scan on "*VALUES*"
+                           Output: "*VALUES*".column1
  Optimizer: legacy query optimizer
- Settings: optimizer=off
-(23 rows)
+(25 rows)
 
 select foo1.join_key as foo1_id, foo3.join_key AS foo3_id, bug_field from
   (values (0),(1)) foo1(join_key)