@pengzhout 2019-03-11
> A better plan for
such case is to redistribute outer relation because general-locus
relation is often small and inner relation might be very large.
I don't think this description is convinced, what if general-locus relation is large enough or inner relation is small enough or both they are small, will this fix still produce a better plan?
 
There is no much difference for data movement between redistribute a larger table and gather a larger table, the difference is: after the redistribution of a table, plan on the top is still a parallel one, the hash cost is much cheaper, the join cost is also cheaper, by contrast, gather motion make it a bottleneck.

This fix still hardcode the orders for general locus same as the rules for partitioned locus (a fake cost model), can we dig a little bit deeper on this problem to make the motion choice a real cost model? cdbpath_motion_for_join() provide all available motion paths, let the cost model choose the cheapest path.