@kainwen 2019-03-11
> I don't think this description is convinced, what if general-locus relation is large enough or inner relation is small enough or both they are small, will this fix still produce a better plan?
> 

Both small data, plan does not impact the complexity (think of sorting small data, insert sort is almost as good as qsort). For huge general and small inner, my benchmark shows the running time is almost the same. (I have not done any mathematic complexity analysis, we should do it). So until now, the new plan is not worse in some case and much better in some case. In summary, it is better.

> There is no much difference for data movement between redistribute a larger table and gather a larger table, the difference is: after the redistribution of a table, plan on the top is still a parallel one, the hash cost is much cheaper, the join cost is also cheaper, by contrast, gather motion make it a bottleneck.
> 
> This fix still hardcode the orders for general locus same as the rules for partitioned locus (a fake cost model), can we dig a little bit deeper on this problem to make the motion choice a real cost model? cdbpath_motion_for_join() provide all available motion paths, let the cost model choose the cheapest path.

Yes. We can do it later or even we can learn from ORCA's result node skill.

The intention for this series of PR is to refactor `cdbpath_motion_for_join`. I think the code is chaos and hard to maintain and understand.  So let's keep MVP in mind.